feat(W1): by-value FFI structs + typed `cstr` (raw char*), fix float-field width
Completes the FFI type surface so `wyn add`'d C libraries with struct-by-value and
raw-char* signatures work.

1. **Struct-by-value across FFI.** `extern fn f(v: MyStruct) -> MyStruct` now
   type-checks and lowers correctly: extern_map_type (checker) resolves a
   user-struct-named param/return to its struct Type; extern_c_type (codegen)
   emits the C struct name by value. Verified end-to-end against a real C lib
   (`double vl(Vec2)` / `Vec2 vscale(Vec2,double)` round-trip: 25.0, (6,8)).

2. **Typed `cstr`.** A raw C string type distinct from Wyn's UTF-8 `string`:
   maps to a plain `char*` both ways (no const/owned wrapping), for C funcs that
   take/return `char*`. `string` stays the default bindgen emits (backward compat).

3. **float struct-field width bug (latent, fixed here).** A `float` struct field
   emitted C `float` (32-bit) while every other Wyn `float` is C `double` (64-bit)
   — so a double value assigned into the field was TRUNCATED, and a Wyn struct
   couldn't match a C struct of `double` fields (broke #1's layout). Now float
   fields emit `double`: full precision (3.14159 no longer truncates) AND correct
   by-value FFI layout.

Gate: make 0-warn -> make test 111/0 + bindgen + cpkg + lsp -> examples unchanged
-> gcc-container: struct-by-value round-trip + cstr + float precision all correct.
New tests/regression/test_struct_float_field.wyn.

+++ src/codegen_stmt.c
@@ -2706,7 +2706,11 @@ void codegen_stmt(Stmt* stmt) {
                         } else if (type_name.length == 3 && memcmp(type_name.start, "int", 3) == 0) {
                             c_type = "long long";
                         } else if (type_name.length == 5 && memcmp(type_name.start, "float", 5) == 0) {
-                            c_type = "float";
+                            // Wyn `float` is C `double` (64-bit) everywhere — values,
+                            // params, returns — so a struct field must be `double`
+                            // too. Emitting `float` (32-bit) truncated precision on
+                            // assignment and broke by-value FFI struct layout.
+                            c_type = "double";
                         } else if (type_name.length == 6 && memcmp(type_name.start, "string", 6) == 0) {
                             c_type = "const char*"; // Always use simple strings for now
                         } else if (type_name.length == 3 && memcmp(type_name.start, "str", 3) == 0) {
